+++ src/compiler/cpp_generator.cc
@@ -141,7 +141,7 @@ std::string GetHeaderIncludes(grpc_generator::File* file,
         "grpcpp/support/async_stream.h",
         "grpcpp/support/async_unary_call.h",
         "grpcpp/impl/codegen/client_callback.h",
-        "grpcpp/impl/codegen/client_context.h",
+        "grpcpp/client_context.h",
         "grpcpp/impl/codegen/completion_queue.h",
         "grpcpp/impl/codegen/message_allocator.h",
         "grpcpp/impl/codegen/method_handler.h",
